requirements.txt
python-telegram-bot==13.7
PyPDF2==3.0.0
python-docx==0.8.11
Pillow==9.0.0
